server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sentence-transformers==3.3.1
transformers==4.47.1
torch>=2.0.0
Pillow>=10.0.0
datasets>=3.0.0
numpy>=1.24.0
pydantic>=2.0.0
